pyproject.toml
[tool.ruff]
line-length = 100
target-version = "py310"
include = ["**/*.py"]
exclude = [
  ".git",
  ".mypy_cache",
  ".pytest_cache",
  ".tox",
  ".venv",
  "__pycache__",
  "build",
  "dist",
  "site-packages",
  "venv",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[project]
name = "dataset-collector"
description = "Dataset collection pipelines and tooling."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dynamic = ["version"]
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.32.2,<2.33",
  "boto3>=1.34.0",
  "datasets>=2.20.0,<2.21",
  "pyarrow>=15.0.0",
  "jsonschema>=4.22.0",
  "zstandard>=0.22.0",
  "pypdf>=4.0.0",
  "pdfminer.six>=20221105",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",
  "trafilatura>=1.6.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy>=1.8.0",
  "types-PyYAML",
  "types-requests",
  "types-jsonschema",
  "yamllint",
  "pytest_httpserver",
  "jupyterlab",
  "ipykernel",
  "pre-commit>=3.6.0",
]

observability = [
  "opentelemetry-api>=1.22.0",
  "opentelemetry-sdk>=1.22.0",
  "opentelemetry-exporter-otlp>=1.22.0",
  "prometheus-client>=0.19.0",
]

async = [
  "aiohttp>=3.9.0",
  "httpx>=0.27.0",
]

dedup = [
  "datasketch>=1.6.0",
]

all = [
  "dataset-collector[dev,observability,async,dedup]",
]

[project.scripts]
dc-review = "collector_core.review_queue:main"
dc-catalog = "collector_core.catalog_builder:main"
dc-pipeline = "collector_core.pipeline_cli:main"
dc = "collector_core.dc_cli:main"

[tool.setuptools.dynamic]
version = { attr = "collector_core.__version__.__version__" }

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.package-data]
collector_core = ["py.typed", "schemas/*.json"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
strict_concatenate = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_return_any = true

[[tool.mypy.overrides]]
module = [
  "datasets.*",
  "boto3.*",
  "botocore.*",
  "datasketch.*",
  "trafilatura.*",
  "pypdf.*",
  "pdfminer.*",
  "bs4.*",
  "lxml.*",
]
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
